test for create mne stuffs, plus small refactoring
